fix: move check into before update entity hook
