+++ apps/hash-api/src/graphql/resolvers/knowledge/entity/entity.ts
@@ -4,8 +4,6 @@ import {
   currentTimeInstantTemporalAxes,
   zeroedGraphResolveDepths,
 } from "@local/hash-isomorphic-utils/graph-queries";
-import { simplifyProperties } from "@local/hash-isomorphic-utils/simplify-properties";
-import { UserProperties } from "@local/hash-isomorphic-utils/system-types/shared";
 import {
   AccountGroupId,
   AccountId,
@@ -41,8 +39,6 @@ import {
   isEntityLinkEntity,
   updateLinkEntity,
 } from "../../../../graph/knowledge/primitive/link-entity";
-import { modifyWebAuthorizationRelationships } from "../../../../graph/ontology/primitive/util";
-import { systemAccountId } from "../../../../graph/system-account";
 import {
   AccountGroupAuthorizationSubjectRelation,
   AuthorizationSubjectKind,
@@ -66,7 +62,6 @@ import {
   ResolverFn,
 } from "../../../api-types.gen";
 import { GraphQLContext, LoggedInGraphQLContext } from "../../../context";
-import { userHasAccessToHash } from "../../shared/user-has-access-to-hash";
 import { dataSourcesToImpureGraphContext } from "../../util";
 import { mapEntityToGQL } from "../graphql-mapping";
 import { createSubgraphAndPermissionsReturn } from "../shared/create-subgraph-and-permissions-return";
@@ -327,60 +322,6 @@ export const updateEntityResolver: ResolverFn<
   });
 
   let updatedEntity: Entity;
-
-  const { shortname, preferredName } = simplifyProperties(
-    updatedProperties as UserProperties,
-  );
-
-  if (isIncompleteUser && shortname && preferredName) {
-    /**
-     * If the user doesn't have access to the HASH instance,
-     * we need to forbid them from completing account signup
-     * and prevent them from receiving ownership of the web.
-     */
-    if (!userHasAccessToHash({ user })) {
-      throw new ForbiddenError(
-        "The user does not have access to the HASH instance, and therefore cannot complete account signup.",
-      );
-    }
-
-    // Now that the user has completed signup, we can transfer the ownership of the web
-    // allowing them to create entities and types.
-    await modifyWebAuthorizationRelationships(
-      context,
-      { actorId: systemAccountId },
-      [
-        {
-          operation: "delete",
-          relationship: {
-            subject: {
-              kind: "account",
-              subjectId: systemAccountId,
-            },
-            resource: {
-              kind: "web",
-              resourceId: user.accountId as OwnedById,
-            },
-            relation: "owner",
-          },
-        },
-        {
-          operation: "create",
-          relationship: {
-            subject: {
-              kind: "account",
-              subjectId: user.accountId,
-            },
-            resource: {
-              kind: "web",
-              resourceId: user.accountId as OwnedById,
-            },
-            relation: "owner",
-          },
-        },
-      ],
-    );
-  }
 
   if (isEntityLinkEntity(entity)) {
     updatedEntity = await updateLinkEntity(context, authentication, {
