+++ apps/hash-api/src/graphql/resolvers/knowledge/user/has-access-to-hash.ts
@@ -1,6 +1,6 @@
 import { Query, ResolverFn } from "../../../api-types.gen";
 import { LoggedInGraphQLContext } from "../../../context";
-import { userHasAccessToHash } from "../../shared/user-has-access-to-hash";
+import { userHasAccessToHash } from "../../../../shared/user-has-access-to-hash";
 
 export const hasAccessToHashResolver: ResolverFn<
   Query["hasAccessToHash"],
